+++ pm/CHANGELOG.md
@@ -23,6 +23,7 @@ description: A detailed, chronological record of all changes made to a project, 
 - Allowed view chat history for deleted characters
 - Fixed chat info query invalidation using locale
 - Fixed chat navigation in approval "message" action
+- Fixed characrer display and avatar name on "Contact Info"
 
 ## [0.1.0-7] - 2025-04-01